WIP (Move to macros::reflection module)
