+++ juniper_codegen/src/graphql_union/mod.rs
@@ -621,30 +621,30 @@ impl Definition {
 
         quote! {
             #[automatically_derived]
-            impl#impl_generics ::juniper::macros::helper::BaseType<#scalar>
+            impl#impl_generics ::juniper::macros::reflection::BaseType<#scalar>
                 for #ty
                 #where_clause
             {
-                const NAME: ::juniper::macros::helper::Type = #name;
+                const NAME: ::juniper::macros::reflection::Type = #name;
             }
 
             #[automatically_derived]
-            impl#impl_generics ::juniper::macros::helper::BaseSubTypes<#scalar>
+            impl#impl_generics ::juniper::macros::reflection::BaseSubTypes<#scalar>
                 for #ty
                 #where_clause
             {
-                const NAMES: ::juniper::macros::helper::Types = &[
-                    <Self as ::juniper::macros::helper::BaseType<#scalar>>::NAME,
-                    #(<#variants as ::juniper::macros::helper::BaseType<#scalar>>::NAME),*
+                const NAMES: ::juniper::macros::reflection::Types = &[
+                    <Self as ::juniper::macros::reflection::BaseType<#scalar>>::NAME,
+                    #(<#variants as ::juniper::macros::reflection::BaseType<#scalar>>::NAME),*
                 ];
             }
 
             #[automatically_derived]
-            impl#impl_generics ::juniper::macros::helper::WrappedType<#scalar>
+            impl#impl_generics ::juniper::macros::reflection::WrappedType<#scalar>
                 for #ty
                 #where_clause
             {
-                const VALUE: ::juniper::macros::helper::WrappedValue = 1;
+                const VALUE: ::juniper::macros::reflection::WrappedValue = 1;
             }
         }
     }
